pyproject.toml
[project]
name = "PriceSentry"
version = "1.0.0"
description = "A lightweight cryptocurrency contract price monitoring tool built for traders and enthusiasts"
readme = "README.md"
requires-python = ">=3.8"
license = { file = "LICENSE" }
authors = [
  { name = "Xeron", email = "xeron2333@gmail.com" },
]
dependencies = [
    "ccxt>=4.1.80",
    "pyyaml>=6.0.1",
    "expiringdict>=1.2.1",
    "pytz>=2024.1",
    "websockets>=11.0.3",
    "matplotlib>=3.7.0",
    "mplfinance>=0.12.9",
    "pandas>=1.5.0",
    "pytest-cov>=5.0.0",
    "psutil>=7.0.0",
    "pytest-asyncio>=0.24.0",
    "fastapi>=0.104.1",
    "uvicorn[standard]>=0.24.0",
    "python-multipart>=0.0.6",
    "ruff>=0.12.8",
    "websocket-client>=1.8.0",
]

[project.urls]
Homepage = "https://github.com/Xeron2000/PriceSentry"
Issues = "https://github.com/Xeron2000/PriceSentry/issues"

[project.optional-dependencies]
dev = [
    "ruff",
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.11.1",
    "pytest-asyncio>=0.21.1",
    "pip-audit>=2.6.0",
    "bandit>=1.7.5",
    "safety>=3.0.0"
]

[tool.ruff]
line-length = 88

[tool.ruff.lint]
select = ["E", "F", "W", "I"]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
include = ["core", "exchanges", "notifications", "utils", "app"]

[dependency-groups]
dev = [
    "ruff>=0.6.0",
    "bandit>=1.7.5",
    "safety>=3.0.0",
    "pip-audit>=2.7.3",
    "typer<0.13.0",
    "mypy>=1.14.1",
]
